refactor: 동아리 모집 공고 관련 서비스 단위 테스트 분리
